blockchain_messenger/requirements.txt
bcrypt==3.2.2
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Classful==0.14.2
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
p2pnetwork==1.2
Pillow==9.1.0
pycparser==2.21
pycryptodomex==3.14.1
six==1.16.0
SQLAlchemy==1.4.36
uuid==1.30
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0